+++ jvm/src/main/scala/gpdviz/ApiImpl.scala
@@ -1,0 +1,9 @@
+package gpdviz
+
+import gpdviz.config.cfg
+
+object ApiImpl extends Api {
+  def usingPusher(): Boolean = cfg.pusher.isDefined
+
+}
+