+++ jvm/src/test/scala/gpdviz/GpdvizSpec.scala
@@ -3,8 +3,7 @@ package gpdviz
 import akka.http.scaladsl.model.ContentTypes._
 import akka.http.scaladsl.model.StatusCodes._
 import akka.http.scaladsl.testkit.ScalatestRouteTest
-import gpdviz.async.PusherNotifier
-import gpdviz.config.cfg
+import gpdviz.async.{Notifier, NullNotifier}
 import gpdviz.data.FileDb
 import gpdviz.model._
 import gpdviz.server._
@@ -13,7 +12,7 @@ import spray.json.JsObject
 
 
 class GpdvizSpec extends WordSpec with Matchers with ScalatestRouteTest with GpdvizService {
-  val notifier = new PusherNotifier(cfg.pusher.get)
+  val notifier: Notifier = NullNotifier
 
   val db = new FileDb("data_test")