+++ shared/src/main/scala/gpdviz/Api.scala
@@ -1,0 +1,7 @@
+package gpdviz
+
+trait Api {
+
+  def usingPusher(): Boolean
+
+}